+++ app/modules/editor/redux/actions/composed/changeTimeItem.ts
@@ -1,7 +1,7 @@
 import { StoreThunkAction } from '../../../../../redux/types';
 import { ImmutableJSTableSetType } from '../../../editorTypes';
 import updateSetDurationForKey from '../../../pure/sets/updateSetDurationForKey';
-import { updateTableDurationBySets } from '../../updateTableDurationBySets';
+import { updateTableDurationBySets } from './updateTableDurationBySets';
 import replaceEditorSets from '../replaceEditorSets';
 
 const changeTimeItem =
